requirements.txt
backports.functools-lru-cache==1.4
certifi==2017.11.5
chardet==3.0.4
click==6.7
cycler==0.10.0
future==0.16.0
idna==2.6
matplotlib==2.1.0
nltk==3.2.5
numpy==1.13.3
oauthlib==2.0.6
pafy==0.5.3.1
pycorpora==0.1.2
pyparsing==2.2.0
python-dateutil==2.6.1
python-dotenv==0.7.1
python-twitter==3.3
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
subprocess32==3.2.7
textblob==0.13.0
urllib3==1.22
youtube-dl==2017.11.6
